requirements.txt
psycopg2
spotipy
youtube-python